device info: Allow zone sizes 4GB and larger on 32-bit systems
zone_size is defined as size_t.
size_t is only 64-bit on 64-bit systems.
Change zone_size data type to unsigned long long in order
to allow zone sizes 4GB and larger also on 32-bit systems.

+++ include/libzbd/zbd.h
@@ -184,7 +184,7 @@ struct zbd_info {
 	/**
 	 * Size in bytes of a zone.
 	 */
-	size_t			zone_size;
+	unsigned long long	zone_size;
 
 	/**
 	 * Size in 512B sectors of a zone.
